requirements.txt
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.2.1
et-xmlfile==1.1.0
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
openpyxl==3.0.10
orjson==3.6.8
packaging==21.3
pandas==1.4.2
psycopg2-binary==2.9.3
pydantic==1.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
redis==4.3.1
requests==2.27.1
rq==1.10.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.19.1
typing-extensions==4.2.0
ujson==5.3.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
wrapt==1.14.1
xlrd==2.0.1
